requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
opencv-python
mss
pytesseract
numpy